move helper function to apollodao index.js
